# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AudioMix": {
        "additionalProperties": false,
        "description": "The two gains a version 1 composition can set.",
        "properties": {
          "gainDb": {
            "maximum": 12.0,
            "minimum": -60.0,
            "title": "Gaindb",
            "type": "number"
          },
          "musicGainDb": {
            "maximum": 12.0,
            "minimum": -60.0,
            "title": "Musicgaindb",
            "type": "number"
          }
        },
        "required": [
          "gainDb",
          "musicGainDb"
        ],
        "title": "AudioMix",
        "type": "object"
      },
...
      "Bookmark": {
        "additionalProperties": false,
        "description": "A member's marker on the timeline.",
        "properties": {
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "timelineMs": {
            "maximum": 600000.0,
            "minimum": 0.0,
            "title": "Timelinems",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "timelineMs",
          "label"
        ],
        "title": "Bookmark",
        "type": "object"
      },
...
      "BrandViolationResponse": {
        "additionalProperties": false,
        "description": "One rule this clip breaks, named where a member can act on it.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "elementId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elementid"
          }
        },
        "required": [
          "code",
          "elementId",
          "detail"
        ],
        "title": "BrandViolationResponse",
        "type": "object"
      },
...
      "BrollDecisionBody": {
        "additionalProperties": false,
        "description": "What one member decided about one proposed picture.",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/BrollDecision"
          },
          "assetId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assetid"
          },
          "suggestionId": {
            "format": "uuid",
            "title": "Suggestionid",
            "type": "string"
          }
        },
        "required": [
          "suggestionId",
          "action"
        ],
        "title": "BrollDecisionBody",
        "type": "object"
      },
      "BrollDecisionRequest": {
        "additionalProperties": false,
        "description": "One decision and the composition it produced, offered as the next Revision.",
        "properties": {
          "composition": {
            "additionalProperties": true,
            "title": "Composition",
            "type": "object"
          },
          "decision": {
            "$ref": "#/components/schemas/BrollDecisionBody"
          },
          "expectedRevision": {
            "minimum": 1.0,
            "title": "Expectedrevision",
            "type": "integer"
          }
        },
        "required": [
          "expectedRevision",
          "composition",
          "decision"
        ],
        "title": "BrollDecisionRequest",
        "type": "object"
      },
...
      "CompositionV1": {
        "additionalProperties": false,
        "description": "One immutable editing decision set, valid on its own terms.",
        "properties": {
          "audio": {
            "$ref": "#/components/schemas/AudioMix"
          },
          "bookmarks": {
            "items": {
              "$ref": "#/components/schemas/Bookmark"
            },
            "title": "Bookmarks",
            "type": "array"
          },
          "brandKit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandKitReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "canvas": {
            "$ref": "#/components/schemas/Canvas"
          },
          "captions": {
            "$ref": "#/components/schemas/Captions"
          },
          "durationMs": {
            "exclusiveMinimum": 0.0,
            "maximum": 600000.0,
            "title": "Durationms",
            "type": "integer"
          },
          "overlays": {
            "items": {
              "discriminator": {
                "mapping": {
                  "citation": "#/components/schemas/CitationOverlay",
                  "image": "#/components/schemas/ImageOverlay",
                  "text": "#/components/schemas/TextOverlay",
                  "video": "#/components/schemas/VideoOverlay"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/VideoOverlay"
                },
                {
                  "$ref": "#/components/schemas/ImageOverlay"
                },
                {
                  "$ref": "#/components/schemas/TextOverlay"
                },
                {
                  "$ref": "#/components/schemas/CitationOverlay"
                }
              ]
            },
            "title": "Overlays",
            "type": "array"
          },
          "schemaVersion": {
            "const": 1,
            "title": "Schemaversion",
            "type": "integer"
          },
          "sourceAssetId": {
            "format": "uuid",
            "title": "Sourceassetid",
            "type": "string"
          },
          "sourceRange": {
            "$ref": "#/components/schemas/SourceRange"
          },
          "template": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TemplateReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/Track"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "schemaVersion",
          "sourceAssetId",
          "durationMs",
          "canvas",
          "sourceRange",
          "template",
          "brandKit",
          "tracks",
          "captions",
          "overlays",
          "audio",
          "bookmarks"
        ],
        "title": "CompositionV1",
        "type": "object"
      },
...
      "EditResponse": {
        "additionalProperties": false,
        "description": "One Edit and the composition its current Revision holds.",
        "properties": {
          "brandViolations": {
            "items": {
              "$ref": "#/components/schemas/BrandViolationResponse"
            },
            "title": "Brandviolations",
            "type": "array"
          },
          "candidateId": {
            "format": "uuid",
            "title": "Candidateid",
            "type": "string"
          },
          "composition": {
            "$ref": "#/components/schemas/CompositionV1"
          },
          "compositionHash": {
            "title": "Compositionhash",
            "type": "string"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "currentRevision": {
            "title": "Currentrevision",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "projectId": {
            "format": "uuid",
            "title": "Projectid",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "candidateId",
          "currentRevision",
          "composition",
          "compositionHash",
          "brandViolations",
          "createdAt",
          "updatedAt"
        ],
        "title": "EditResponse",
        "type": "object"
      },
...
      "EditSelectionRequest": {
        "additionalProperties": false,
        "description": "The look and the brand a member chose when they opened this clip.",
        "properties": {
          "brandKitId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brandkitid"
          },
          "templateId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Templateid"
          }
        },
        "title": "EditSelectionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RevisionHistoryResponse": {
        "additionalProperties": false,
        "description": "One Edit's Revisions, newest first.",
        "properties": {
          "revisions": {
            "items": {
              "$ref": "#/components/schemas/RevisionResponse"
            },
            "title": "Revisions",
            "type": "array"
          }
        },
        "required": [
          "revisions"
        ],
        "title": "RevisionHistoryResponse",
        "type": "object"
      },
      "RevisionResponse": {
        "additionalProperties": false,
        "description": "One entry of an Edit's history, without repeating its whole composition.",
        "properties": {
          "compositionHash": {
            "title": "Compositionhash",
            "type": "string"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "format": "uuid",
            "title": "Createdby",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "revision",
          "compositionHash",
          "createdBy",
          "createdAt"
        ],
        "title": "RevisionResponse",
        "type": "object"
      },
      "SaveRevisionRequest": {
        "additionalProperties": false,
        "description": "One composition offered as the successor of the Revision the caller holds.",
        "properties": {
          "composition": {
            "additionalProperties": true,
            "title": "Composition",
            "type": "object"
          },
          "expectedRevision": {
            "minimum": 1.0,
            "title": "Expectedrevision",
            "type": "integer"
          }
        },
        "required": [
          "expectedRevision",
          "composition"
        ],
        "title": "SaveRevisionRequest",
        "type": "object"
      },
...
      "TemplateReference": {
        "additionalProperties": false,
        "description": "The exact template version this composition was built against.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "version": {
            "exclusiveMinimum": 0.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version"
        ],
        "title": "TemplateReference",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Clipah API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/edits/{edit_id}": {
      "get": {
        "description": "Read one Edit, hiding another Workspace's Edit behind the same absence.",
        "operationId": "show_api_v1_edits__edit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Show",
        "tags": [
          "edits"
        ]
      },
      "put": {
        "description": "Append one composition as the next Revision, or refuse a stale expectation.",
        "operationId": "save_api_v1_edits__edit_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save",
        "tags": [
          "edits"
        ]
      }
    },
...
    "/api/v1/edits/{edit_id}/broll-decisions": {
      "post": {
        "description": "Record one B-roll decision and the Revision it produced, in one transaction.",
        "operationId": "decide_api_v1_edits__edit_id__broll_decisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrollDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide",
        "tags": [
          "edits"
        ]
      }
    },
...
    "/api/v1/edits/{edit_id}/revisions": {
      "get": {
        "description": "List one Edit's Revisions, newest first.",
        "operationId": "history_api_v1_edits__edit_id__revisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History",
        "tags": [
          "edits"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/candidates/{candidate_id}/edits": {
      "post": {
        "description": "Open the one Edit belonging to a reviewed candidate, or reach the existing one.",
        "operationId": "create_api_v1_projects__project_id__candidates__candidate_id__edits_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/EditSelectionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Selection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "edits"
        ]
      }
    },
...
  }
}